created directory for shared components, moved navbar location
